refactored SessionRefresher and added documentation in comments to make it clear that it is an example for a single client
